Broke out the logic in the CallGraph file into several files.
