+++ src/QsCompiler/Transformations/CallGraph/CallGraphWalker.cs
